requirements.txt
asgiref==3.8.1
Django==4.2.16
django-multiselectfield==1.0.1
django-recaptcha==4.0.0
sqlparse==0.5.1

certifi==2024.7.4
charset-normalizer==3.4.0
idna==3.10
pytz==2025.1
requests==2.32.3
urllib3==2.2.3

psycopg2-binary==2.9.10
Pillow==11.1.0
django-imagekit==5.0.0
python-dotenv==1.0.1
